requirements.txt
# Dépendances principales
requests>=2.23.0,<3.0.0
gitpython>=3.1.2,<4.0.0
attrs>=20.1.0,<21.0.0
cattrs>=1.1.2,<2.0.0
pendulum>=2.1.0,<3.0.0
click>=7.1.2,<8.0.0
tqdm>=4.46.0,<5.0.0
matplotlib>=3.2.1,<4.0.0
jinja2>=3.1.4,<4.0.0
pyyaml==5.3.1
graphviz>=0.14.1,<1.0.0
ansible>=4.2.0,<5.0.0

# Dépendances de développement
pytest>=5.4.1,<6.0.0
pytest-cov>=2.8.1,<3.0.0
flake8>=3.8.0-alpha.2,<4.0.0  # Inclut prereleases
setuptools>=46.1.3,<60.0.0
pytest-datadir>=1.3.1,<2.0.0
betamax>=0.8.1,<1.0.0
pytest-flake8>=1.0.5,<2.0.0
pytest-mypy>=0.6.2,<1.0.0
coverage>=5.1,<6.0.0
pytest-sugar>=0.9.3,<1.0.0
mypy>=0.770,<1.0.0
data-science-types>=0.2.10,<1.0.0
wemake-python-styleguide>=0.14.1,<1.0.0
flake8-pytest-style>=1.3.0,<2.0.0
flake8-annotations>=2.4.1,<3.0.0
flake8-logging-format>=0.6.0,<1.0.0
docutils>=0.20,<1.0.0
